backend/requirements.txt
Flask
Flask_Cors
python-dotenv
chromadb
tiktoken
pymongo
pinecone-client
langchain
openai
jq
Flask-PyMongo